feat: add ticket type selection
