--- src/components/Dashboard/SectionSubtitle.js
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export default styled.div`
-  font-size: 20px;
-  color: #8E8E8E;
-  margin-top: 30px;
-  margin-bottom: 15px;
-
-  @media (max-width: 600px) {
-    font-size: 16px;
-    margin-top: 15px;
-    margin-bottom: 8px;
-  }
-`;
