--- src/components/Dashboard/SectionTitle.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export default styled.div`
-  font-size: 34px;
-  @media (max-width: 600px) {
-    font-size: 26px;
-  }
-`;
