+++ src/components/PaymentForm/index.js
@@ -1,0 +1,5 @@
+import SelectTicket from "./TicketSelection";
+
+export default function PaymentForm() {
+  return (<SelectTicket/>);
+}
